Split into Modules
- Table class is moved into the module `monsql/table.py`.
- Configuration values such as the TRANSACTION_MODE are placed in
`monsql/config.py`
- `monsql/wrapper.py` holds the wrapper representing the DB object
