+++ monsql/config.py
@@ -2,3 +2,9 @@
 
 ASCENDING = 1
 DESCENDING = 2
+
+
+class TRANSACTION_MODE:
+    AUTO = "auto"
+    MANUAL = "manual"
+    DEFAULT = "manual"
